fix: Biquad 高通系数修正（RBJ 公式 b0/b1/b2 此前写错成直通形状致 8 倍增益）——上轮提交在测试未绿时发出（纪律违规，本轮测试输出确认绿后才提交）

+++ app/src/test/java/com/guitarcoach/app/core/audio/StereoFocusProcessorTest.kt
@@ -68,7 +68,7 @@ class StereoFocusProcessorTest {
     @Test
     fun `单声道降级路径仅带通不崩`() {
         val proc = StereoFocusProcessor(22050)
-        val mono = ShortArray(4410) { i -> (sin(2 * PI * 200 * i / 22050) * 12000).toShort() }
+        val mono = ShortArray(4410) { i -> (sin(2 * PI * 200 * i / 22050) * 12000).toInt().toShort() }
         val out = proc.process(mono, 1)
         assertEquals(mono.size, out.size)
     }
